Marcos: Organizando os arquivos dentro da pasta src, Incompleto ainda é preciso implementar as rotas de maneira correta
