+++ lib/src/services/auth_service.dart
@@ -1,5 +1,5 @@
 class AuthService {
-  static bool isLoggedIn = false; // false para ter que logar
+  static bool isLoggedIn = true; // false para ter que logar
 
   static void login(){
     isLoggedIn = true;
