--- lib/views/Principal.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/material.dart';
-
-class Principal extends StatefulWidget {
-  const Principal({super.key});
-
-  @override
-  State<Principal> createState() => _PrincipalState();
-}
-
-class _PrincipalState extends State<Principal> {
-  @override
-  Widget build(BuildContext context) {
-    return const Placeholder();
-  }
-}
